+++ src/lib/gsvg.js
@@ -8,6 +8,7 @@ const NS           = 'http://www.w3.org/2000/svg';
 const SVG          = 'svg';
 const PATH         = 'path';
 const D            = 'd';
+const TRANSFORM    = 'transform';
 const cache        = new WeakMap();
 const readonlyProp = new Set();
 
@@ -17,7 +18,7 @@ const readonlyProp = new Set();
  * @param {string} prop
  * @returns {boolean}
  */
-const directAccess = (prop) => prop[0] === '_' || isSymbol(prop) || ['el','gSVG'].includes(prop);
+const directAccess = (prop) => prop[0] === '_' || isSymbol(prop) || ['el', 'gSVG'].includes(prop);
 
 /**
  * Check the instance
@@ -199,20 +200,24 @@ const wrapper = (element) => {
             return wrapped[prop].call(proxy, ...args);
           }
         }
-        // Special case <path d=""/>
-        if (prop === D && element.tagName.toLowerCase() === PATH) {
+        // Special cases path d="" and transform=""
+        if (
+          (prop === D && element.tagName.toLowerCase() === PATH) ||
+          prop === TRANSFORM
+        ) {
           let content  = element.getAttribute(D) || '';
+          const processor = prop === D ? pathD : elTransform;
           const dProxy = new Proxy(
             (arg) => {
               preCall(proxy, prop, [arg])
               return isString(arg) ?
-                element.setAttribute(D, arg) || proxy :
-                element.getAttribute(D)
+                element.setAttribute(prop, arg) || proxy :
+                element.getAttribute(prop)
             },
             {
               get (_target, command) {
                 return (...args) => {
-                  content += pathD(proxy, command, args);
+                  content += processor(proxy, command, args);
                   element?.setAttribute(prop, content);
                   return dProxy;
                 };
@@ -221,19 +226,19 @@ const wrapper = (element) => {
           );
           return dProxy;
         }
-        // Special case <path g-bind:d=""/>
-        if (prop === '$d') {
+        // Special cases g-bind:d="" and g-bind:transport=""
+        if (prop === '$d' || prop === '$transform') {
           let content  = '';
+          const processor = prop === '$d' ? pathD : elTransform;
           const dProxy = new Proxy(
-						{},
+            {},
             {
               get (_target, command) {
                 return (...args) => {
-                  if (command === Symbol.toPrimitive)  {
+                  if (command === Symbol.toPrimitive) {
                     return content
                   }
-                  const d = pathD(proxy, command, args);
-                  content += d;
+                  content += processor(proxy, command, args);
                   return dProxy;
                 };
               }
@@ -310,7 +315,9 @@ const methodWrapper = (element, prop, parentWrapper, parentProp) => {
     if (value !== 0 && !value) {
       element?.removeAttribute && element.removeAttribute(propNormalized);
     } else {
-      element?.setAttribute && element.setAttribute(propNormalized, isBoolean(value) ? '' : String(args));
+      element?.setAttribute && element.setAttribute(propNormalized, isBoolean(value) ?
+        '' :
+        String(args));
     }
     return parentWrapper;
   };
@@ -386,6 +393,17 @@ const pathD = (wrapped, prop, args) => {
 }
 
 /**
+ * elTransform
+ * @param {Object} wrapped
+ * @param {string} prop
+ * @param {Array} args
+ */
+const elTransform = (wrapped, prop, args) => {
+  preCall(wrapped, `transform.${ prop }`, args);
+  return `${ prop }(${ args.join(COMA) })`
+}
+
+/**
  * @typedef {function} gSVG
  */
 
@@ -419,8 +437,7 @@ gSVG.extend = (plugin) => {
 }
 
 const setup = {
-  install      : install,
-  installAsync : installAsync,
+  install : install,
   extendConstructor (extension) {
     Object.assign(gSVG, extension);
   },
@@ -442,24 +459,18 @@ const setup = {
 
 /**
  * gSVG.install - load new plugins
- * @param {Function|string} plugin
+ * @param {Function|{svg: Function}} plugin
  * @return {gSVG|Promise}
  */
 function install (plugin) {
-  if (isString(plugin)) {
-    return installAsync(plugin);
+  if (isFunction(plugin)) {
+    plugin(setup);
+  } else if (isFunction(plugin?.svg)) {
+    plugin.svg(setup);
   }
-  plugin(setup);
   return gSVG;
 }
 
 gSVG.install = install;
-
-async function installAsync (source) {
-  const plugin = (await import(source)).default;
-  return install(plugin)
-}
-
-gSVG.installAsync = installAsync;
 
 export default gSVG;