+++ src/lib/plugins/gsvg.shapes.script.js
@@ -1,4 +1,0 @@
-import plugin from './gsvg.shapes.js';
-if (gSVG) {
-  gSVG.install(plugin);
-}