+++ src/plugins/animateto.script.js
@@ -1,0 +1,4 @@
+import {svgPlugin} from './animateto.js';
+if (gSVG) {
+  gSVG.install(svgPlugin);
+}