+++ test/component/template/cases/case08.js
@@ -1,5 +1,0 @@
-export const title       = '8) Handle events';
-export const description = 'You can capture with <code>addEventListener</code> the regular events for Graphery Components and HTMLElements.';
-export const events      = true;
-
-export default '<g-template svg-src="/test/component/template/assets/check.svg" style="width: 50px;"></g-template>';